# backend/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Reliabilix GreenOps API
  version: 1.0.0
  description: Multi-tenant FinOps + GreenOps API. Tenancy comes from the verified identity (membership in the database) or from an API key, never
    from request input. Every tenant table is protected by Row Level Security.
servers:
- url: https://api.reliabilix.green
security:
- bearerAuth: []
...
paths:
...
  /api/v1/finops/summary:
    get:
      operationId: getFinopsSummary
      summary: Spend by service and the preceding period
      tags:
      - finops
      description: Requires permission `usage:read`.
      parameters:
      - $ref: '#/components/parameters/from'
      - $ref: '#/components/parameters/to'
      - $ref: '#/components/parameters/project_id'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinopsSummary'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
  /api/v1/finops/budgets:
    get:
      operationId: listBudgets
      summary: Budgets with current-period consumption
      tags:
      - finops
      description: Requires permission `billing:read`.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/BudgetStatus'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
    post:
      operationId: createBudget
      summary: Create a budget
      tags:
      - finops
      description: Requires permission `budget:write`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BudgetInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Budget'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
  /api/v1/finops/budgets/{id}:
    delete:
      operationId: deleteBudget
      summary: Delete a budget
      tags:
      - finops
      description: Requires permission `budget:write`.
      parameters:
      - $ref: '#/components/parameters/id'
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
  /api/v1/finops/anomalies:
    get:
      operationId: listAnomalies
      summary: Days whose cost is > 3 standard deviations above the trailing mean
      tags:
      - finops
      description: Requires permission `billing:read`.
      parameters:
      - $ref: '#/components/parameters/from'
      - $ref: '#/components/parameters/to'
      - $ref: '#/components/parameters/project_id'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Anomaly'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
  /api/v1/dashboard/trend:
    get:
      operationId: dashboardTrend
      summary: Cost and CO2e per day
      tags:
      - dashboard
      description: Requires permission `usage:read`.
      parameters:
      - $ref: '#/components/parameters/from'
      - $ref: '#/components/parameters/to'
      - $ref: '#/components/parameters/project_id'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/DashboardTrendPoint'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
  /api/v1/dashboard/providers:
    get:
      operationId: dashboardProviders
      summary: CO2e by cloud provider (only providers with data)
      tags:
      - dashboard
      description: Requires permission `usage:read`.
      parameters:
      - $ref: '#/components/parameters/from'
      - $ref: '#/components/parameters/to'
      - $ref: '#/components/parameters/project_id'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProviderShare'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
  /api/v1/dashboard/services:
    get:
      operationId: dashboardServices
      summary: Cost and CO2e by service
      tags:
      - dashboard
      description: Requires permission `usage:read`.
      parameters:
      - $ref: '#/components/parameters/from'
      - $ref: '#/components/parameters/to'
      - $ref: '#/components/parameters/project_id'
      - name: category
        in: query
        schema:
          type: string
          enum:
          - compute
          - storage
          - database
          - networking
          - other
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ServiceRow'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
  /api/v1/dashboard/regions:
    get:
      operationId: dashboardRegions
      summary: Grid carbon intensity by region (reference data)
      tags:
      - dashboard
      description: Requires permission `usage:read`.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/RegionIntensity'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
  /api/v1/dashboard/activity:
    get:
      operationId: dashboardActivity
      summary: Recent user-facing activity
      tags:
      - dashboard
      description: Requires permission `usage:read`.
      parameters:
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 50
          default: 10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Activity'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Auth0 access token (RS256) or an API key (`grk_...`). Optional header `X-Tenant-ID` selects among several memberships.
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    from:
      name: from
      in: query
      schema:
        type: string
        format: date
      description: 'Inclusive start (default: 30 days ago)'
    to:
      name: to
      in: query
      schema:
        type: string
        format: date
      description: 'Inclusive end (default: today)'
    project_id:
      name: project_id
      in: query
      schema:
        type: string
        format: uuid
  responses:
    Problem:
      description: RFC 9457 problem details
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      properties:
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        request_id:
          type: string
      required:
      - title
      - status
      additionalProperties: true
...
    FinopsSummary:
      type: object
      properties:
        has_data:
          type: boolean
        total_cost:
          type: number
        previous_total_cost:
          type: number
        currency:
          type: string
        by_service:
          type: array
          items:
            type: object
            properties:
              service:
                type: string
              cost:
                type: number
            additionalProperties: true
        previous_comparable:
          type: boolean
          description: False when the current or previous window has data for fewer than 80% of its days; clients must not show a delta then.
      additionalProperties: true
    Budget:
      type: object
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
          nullable: true
        amount:
          type: number
        currency:
          type: string
        period:
          type: string
          enum:
          - monthly
          - quarterly
          - yearly
        created_at:
          type: string
          format: date-time
      additionalProperties: true
    BudgetInput:
      type: object
      properties:
        project_id:
          type: string
          format: uuid
          nullable: true
        amount:
          type: number
          minimum: 0
        currency:
          type: string
          minLength: 3
          maxLength: 3
        period:
          type: string
          enum:
          - monthly
          - quarterly
          - yearly
      required:
      - amount
      - currency
      - period
      additionalProperties: true
    BudgetStatus:
      allOf:
      - $ref: '#/components/schemas/Budget'
      - type: object
        properties:
          period_start:
            type: string
            format: date-time
          period_end:
            type: string
            format: date-time
          spent:
            type: number
          used_pct:
            type: number
          projected:
            type: number
          state:
            type: string
            enum:
            - ok
            - at_risk
            - exceeded
        additionalProperties: true
    Anomaly:
      type: object
      properties:
        day:
          type: string
          format: date-time
        cost:
          type: number
        expected:
          type: number
        stddev:
          type: number
      additionalProperties: true
    DashboardTrendPoint:
      type: object
      properties:
        day:
          type: string
          format: date-time
        cost:
          type: number
        carbon_kg_co2e:
          type: number
      additionalProperties: true
    ProviderShare:
      type: object
      properties:
        provider:
          type: string
        carbon_kg_co2e:
          type: number
        share_pct:
          type: number
      additionalProperties: true
    ServiceRow:
      type: object
      properties:
        service:
          type: string
        category:
          type: string
        cost:
          type: number
        carbon_kg_co2e:
          type: number
        trend:
          type: array
          items:
            type: number
      additionalProperties: true
    RegionIntensity:
      type: object
      properties:
        region:
          type: string
        g_per_kwh:
          type: number
        change_pct:
          type: number
          nullable: true
        at:
          type: string
          format: date-time
      additionalProperties: true
    Activity:
      type: object
      properties:
        id:
          type: integer
        action:
          type: string
        target:
          type: string
        at:
          type: string
          format: date-time
      additionalProperties: true
...